// src/shared/enums/ERoutesName.ts
export enum ERoutesName {
  BaseLayout = 'BaseLayout',
  AnalyticsPage = 'AnalyticsPage',
  VacanciesPage = 'VacanciesPage',
  FavouritesPage = 'FavouritesPage',
  VacanciesMapPage = 'VacanciesMapPage',
  VacancyDetailPage = 'VacancyDetailPage',
  NotFoundPage = 'NotFoundPage',
  AuthPage = 'AuthPage',
  AnyPath = 'AnyPath',
}